Perform easing after `FixedMain` loop (#5)
Fixes #2.
A working and up-to-date version of #1.

Currently, easing is performed in `PostUpdate`. This causes major issues if something else depends on the positions of interpolated objects, such as a camera that follows an entity in `Update`. This can be seen below.

https://github.com/user-attachments/assets/e3091d62-de37-4d03-bf62-2b55ab652829

(The frame of reference isn't very clear here, but the camera is following the cyan square, and the red square is stationary)

The problem here is that right after `FixedMain`, the transform of the entity matches the "gameplay transform", i.e. the end state of the interpolation, so the camera following logic in `Update` initially uses that. Then easing is performed in `PostUpdate`, which causes the entities to "snap back" to the start of the easing, making the camera's position incorrect, causing very noticeable visual artifacts.

To fix this frame delay, we should perform the interpolation right after the `FixedMain` loop, in `RunFixedMainLoopSystem::AfterFixedMainLoop`. No more artifacts!

https://github.com/user-attachments/assets/82491956-7204-4221-b3a2-936fcaab37d1

(Again, it looks like the red square is moving, but it's actually the cyan square)